requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
base58==2.1.1
bitarray==2.9.2
certifi==2024.2.2
charset-normalizer==3.3.2
cryptocompare==0.7.6
cytoolz==0.12.3
ecdsa==0.18.0
eth-account==0.11.0
eth-hash==0.6.0
eth-keyfile==0.7.0
eth-keys==0.5.0
eth-rlp==1.0.1
eth-typing==4.0.0
eth-utils==3.0.0
eth_abi==5.0.0
frozenlist==1.4.1
hdwallet==2.2.1
hexbytes==0.3.1
idna==3.6
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lru-dict==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
mnemonic==0.21
multidict==6.0.5
parsimonious==0.9.0
protobuf==4.25.2
pycryptodome==3.20.0
Pygments==2.17.2
pyunormalize==15.1.0
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rich-argparse==1.4.0
rlp==4.0.0
rpds-py==0.17.1
six==1.16.0
tabulate==0.9.0
toolz==0.12.1
typing_extensions==4.9.0
urllib3==2.2.0
web3==6.15.1
websockets==12.0
yarl==1.9.4
